Add option to report relay serial as USB serial
The original behavior of the relays reports a blank USB serial number,
with the only way to get the serial number being through the HID
interface. This is undesirable when multiple boards are connected, as it
makes it hard to distinguish between boards. This optional feature makes
the software report the HID serial number as the USB serial number also,
but comes at the cost of more RAM usage, code space, and deviating from
the original behavior